Refacatored code.
