+++ .gitignore
@@ -1,3 +1,4 @@
 .vscode/
 __pycache__/
 utils/__pycache__/
+.venv/
