+++ docker-setup/.env
@@ -1,0 +1,4 @@
+mysql_host=localhost
+mysql_user=root
+mysql_pass=password
+mysql_db=mini_project
