+++ docker-setup/docker-compose.yml
@@ -1,0 +1,20 @@
+version: "3.8"
+services:
+  db:
+    image: mysql
+    container_name: mysql_container
+    restart: always
+    environment:
+      MYSQL_ROOT_PASSWORD: "${mysql_pass}"
+    ports:
+      - "3306:3306"
+    volumes:
+      - example_db
+  adminer:
+    image: adminer
+    container_name: adminer_container
+    restart: always
+    ports:
+      - 8080:8080
+volumes:
+  example_db:
