+++ docker-setup/requirements copy.txt
@@ -1,0 +1,6 @@
+cffi==1.14.4
+cryptography==3.2.1
+pycparser==2.20
+PyMySQL==0.10.1
+python-dotenv==0.15.0
+six==1.15.0
