--- utils/test/test_courier_list.csv
@@ -1,3 +1,0 @@
-Index,Name,Age,Vehicle,Status
-1,Martin,27,Car,Busy
-2,Sheila,41,Car,Busy
