--- utils/test/test_write_to_csv.csv
@@ -1,3 +1,0 @@
-Name,Age
-Martin,25
-Sheila,30
